[C3] Reorganize helper scripts (#5218)
* C3: Update SolidStart template to work with create-solid@0.5.3

* Move sleep helper to its own file

* Move package management helpers out of command.ts into dedicated files

* Move the retry helper to its own file

* Consolidate compat date helpers in their own file

* Move installWrangler to packages

* Remove logging helper. This problem was solved by the e2e logging rework

* Improve docstrings

* fixup import statements

* Move wrangler.toml related functions into their own file

* Move wrangler accounts functions into dedicated file. Move wrangler folder out of helpers

* Relocate runFrameworkGenerator

* Add unit tests for runFrameworkGenerator

* Move quoteShellArgs to command

* Move validateProjectDirectory to validators helper

* Move git helpers to dedicated file

* Reorganize remaining bits of common.ts

* Add unit tests for git helpers

* Adding unit tests for deploy helpers

* Make spinner mocking consistent across tests

* Remove runCommands and add tests for gitCommit

* Remove the mockRunCommand helper

* Addressing PR feedback

* changeset
