+++ packages/create-cloudflare/e2e-tests/frameworks.test.ts
@@ -1,9 +1,9 @@
 import { existsSync } from "fs";
 import { cp } from "fs/promises";
 import { join } from "path";
-import { retry } from "helpers/command";
-import { sleep } from "helpers/common";
-import { detectPackageManager } from "helpers/packages";
+import { detectPackageManager } from "helpers/packageManagers";
+import { retry } from "helpers/retry";
+import { sleep } from "helpers/sleep";
 import { fetch } from "undici";
 import {
 	afterEach,
