+++ packages/create-cloudflare/e2e-tests/helpers.ts
@@ -10,8 +10,8 @@ import { tmpdir } from "os";
 import { basename, join } from "path";
 import { stripAnsi } from "@cloudflare/cli";
 import { spawn } from "cross-spawn";
-import { retry } from "helpers/command";
-import { sleep } from "helpers/common";
+import { retry } from "helpers/retry";
+import { sleep } from "helpers/sleep";
 import { fetch } from "undici";
 import { expect } from "vitest";
 import { version } from "../package.json";
