+++ packages/create-cloudflare/src/__tests__/workers.test.ts
@@ -1,89 +1,22 @@
-import { existsSync, readdirSync } from "fs";
-import { spinner } from "@cloudflare/cli/interactive";
-import { getWorkerdCompatibilityDate } from "helpers/command";
+import { existsSync } from "fs";
+import { mockSpinner } from "helpers/__tests__/mocks";
+import { getLatestTypesEntrypoint } from "helpers/compatDate";
 import { readFile, writeFile } from "helpers/files";
 import { beforeEach, describe, expect, test, vi } from "vitest";
-import {
-	addWorkersTypesToTsConfig,
-	getLatestTypesEntrypoint,
-	updateWranglerToml,
-} from "../workers";
+import { addWorkersTypesToTsConfig } from "../workers";
 import { createTestContext } from "./helpers";
-import type { Dirent } from "fs";
 import type { C3Context } from "types";
 
 vi.mock("helpers/files");
-vi.mock("helpers/command");
+vi.mock("helpers/compatDate");
 vi.mock("fs");
 vi.mock("@cloudflare/cli/interactive");
 
 beforeEach(() => {
-	// we mock `spinner` to remove noisy logs from the test runs
-	vi.mocked(spinner).mockImplementation(() => ({
-		start() {},
-		update() {},
-		stop() {},
-	}));
+	mockSpinner();
 });
 
-const mockWorkersTypesDirListing = [
-	"2021-11-03",
-	"2022-03-21",
-	"2022-11-30",
-	"2023-03-01",
-	"2023-07-01",
-	"experimental",
-	"index.d.ts",
-	"index.ts",
-	"oldest",
-	"package.json",
-];
-
-const mockWorkersTypesDirectory = (
-	mockImpl: () => string[] = () => [...mockWorkersTypesDirListing]
-) => {
-	vi.mocked(readdirSync).mockImplementation((path) => {
-		if (path.toString().match("workers-types")) {
-			// vitest won't resolve the type for the correct `readdirSync` overload thus the trickery
-			return mockImpl() as unknown as Dirent[];
-		}
-		return [];
-	});
-};
-
-describe("getLatestTypesEntrypoint", () => {
-	const ctx = createTestContext();
-
-	test("happy path", async () => {
-		mockWorkersTypesDirectory();
-
-		const entrypoint = getLatestTypesEntrypoint(ctx);
-		expect(entrypoint).toBe("2023-07-01");
-	});
-
-	test("read error", async () => {
-		mockWorkersTypesDirectory(() => {
-			throw new Error("ENOENT: no such file or directory");
-		});
-
-		const entrypoint = getLatestTypesEntrypoint(ctx);
-		expect(entrypoint).toBe(null);
-	});
-
-	test("empty directory", async () => {
-		mockWorkersTypesDirectory(() => []);
-
-		const entrypoint = getLatestTypesEntrypoint(ctx);
-		expect(entrypoint).toBe(null);
-	});
-
-	test("no compat dates found", async () => {
-		mockWorkersTypesDirectory(() => ["foo", "bar"]);
-
-		const entrypoint = getLatestTypesEntrypoint(ctx);
-		expect(entrypoint).toBe(null);
-	});
-});
+const mockCompatDate = "2024-01-17";
 
 describe("addWorkersTypesToTsConfig", () => {
 	let ctx: C3Context;
@@ -93,7 +26,7 @@ describe("addWorkersTypesToTsConfig", () => {
 		ctx.args.ts = true;
 
 		vi.mocked(existsSync).mockImplementation(() => true);
-		mockWorkersTypesDirectory();
+		vi.mocked(getLatestTypesEntrypoint).mockReturnValue(mockCompatDate);
 
 		// Mock the read of tsconfig.json
 		vi.mocked(readFile).mockImplementation(
@@ -107,7 +40,7 @@ describe("addWorkersTypesToTsConfig", () => {
 		expect(writeFile).toHaveBeenCalled();
 
 		expect(vi.mocked(writeFile).mock.calls[0][1]).toContain(
-			`"@cloudflare/workers-types/2023-07-01"`
+			`"@cloudflare/workers-types/${mockCompatDate}"`
 		);
 	});
 
@@ -118,9 +51,7 @@ describe("addWorkersTypesToTsConfig", () => {
 	});
 
 	test("latest entrypoint not found", async () => {
-		vi.mocked(readdirSync).mockImplementation(
-			() => ["README.md"] as unknown as Dirent[]
-		);
+		vi.mocked(getLatestTypesEntrypoint).mockReturnValue(null);
 		await addWorkersTypesToTsConfig(ctx);
 
 		expect(writeFile).not.toHaveBeenCalled();
@@ -136,94 +67,5 @@ describe("addWorkersTypesToTsConfig", () => {
 		expect(vi.mocked(writeFile).mock.calls[0][1]).toContain(
 			`"@cloudflare/workers-types/2021-03-20"`
 		);
-	});
-});
-
-describe("updateWranglerToml", () => {
-	const ctx = createTestContext();
-
-	const mockCompatDate = "2024-01-17";
-
-	beforeEach(() => {
-		vi.mocked(getWorkerdCompatibilityDate).mockReturnValue(
-			Promise.resolve(mockCompatDate)
-		);
-		vi.mocked(existsSync).mockImplementation(() => true);
-		mockWorkersTypesDirectory();
-
-		// Mock the read of tsconfig.json
-		vi.mocked(readFile).mockImplementation(
-			() => `{ "compilerOptions": { "types": ["@cloudflare/workers-types"]} }`
-		);
-	});
-
-	test("placeholder replacement", async () => {
-		const toml = [
-			`name = "<TBD>"`,
-			`main = "src/index.ts"`,
-			`compatibility_date = "<TBD>"`,
-		].join("\n");
-		vi.mocked(readFile).mockReturnValue(toml);
-
-		await updateWranglerToml(ctx);
-
-		const newToml = vi.mocked(writeFile).mock.calls[0][1];
-		expect(newToml).toMatch(`name = "${ctx.project.name}"`);
-		expect(newToml).toMatch(`main = "src/index.ts"`);
-		expect(newToml).toMatch(`compatibility_date = "${mockCompatDate}"`);
-	});
-
-	test("empty replacement", async () => {
-		const toml = [
-			`name = `,
-			`main = "src/index.ts"`,
-			`compatibility_date = `,
-		].join("\n");
-		vi.mocked(readFile).mockReturnValue(toml);
-
-		await updateWranglerToml(ctx);
-
-		const newToml = vi.mocked(writeFile).mock.calls[0][1];
-		expect(newToml).toMatch(`name = "${ctx.project.name}"`);
-		expect(newToml).toMatch(`main = "src/index.ts"`);
-		expect(newToml).toMatch(`compatibility_date = "${mockCompatDate}"`);
-	});
-
-	test("string literal replacement", async () => {
-		const toml = [`name = "my-cool-worker"`, `main = "src/index.ts"`].join(
-			"\n"
-		);
-		vi.mocked(readFile).mockReturnValue(toml);
-
-		await updateWranglerToml(ctx);
-
-		const newToml = vi.mocked(writeFile).mock.calls[0][1];
-		expect(newToml).toMatch(`name = "${ctx.project.name}"`);
-		expect(newToml).toMatch(`main = "src/index.ts"`);
-	});
-
-	test("missing name and compat date", async () => {
-		const toml = `main = "src/index.ts"`;
-		vi.mocked(readFile).mockReturnValue(toml);
-
-		await updateWranglerToml(ctx);
-
-		const newToml = vi.mocked(writeFile).mock.calls[0][1];
-		expect(newToml).toMatch(`name = "${ctx.project.name}"`);
-		expect(newToml).toMatch(`main = "src/index.ts"`);
-		expect(newToml).toMatch(`compatibility_date = "${mockCompatDate}"`);
-	});
-
-	test("dont replace valid existing compatibility date", async () => {
-		const toml = [
-			`name = "super-old-worker"`,
-			`compatibility_date = "2001-10-12"`,
-		].join("\n");
-		vi.mocked(readFile).mockReturnValue(toml);
-
-		await updateWranglerToml(ctx);
-
-		const newToml = vi.mocked(writeFile).mock.calls[0][1];
-		expect(newToml).toMatch(`compatibility_date = "2001-10-12"`);
 	});
 });
