+++ packages/create-cloudflare/src/helpers/cli.ts
@@ -2,10 +2,10 @@ import { updateStatus, warn } from "@cloudflare/cli";
 import { blue } from "@cloudflare/cli/colors";
 import { spinner, spinnerFrames } from "@cloudflare/cli/interactive";
 import Haikunator from "haikunator";
+import { getLatestPackageVersion } from "helpers/packages";
 import open from "open";
 import semver from "semver";
 import { version } from "../../package.json";
-import { getLatestPackageVersion } from "./latestPackageVersion";
 
 /**
  * An extremely simple wrapper around the open command.
