+++ packages/create-cloudflare/templates/gatsby/c3.ts
@@ -1,6 +1,6 @@
 import { inputPrompt } from "@cloudflare/cli/interactive";
-import { runFrameworkGenerator } from "helpers/command";
-import { detectPackageManager } from "helpers/packages";
+import { runFrameworkGenerator } from "frameworks/index";
+import { detectPackageManager } from "helpers/packageManagers";
 import type { TemplateConfig } from "../../src/templates";
 import type { C3Context } from "types";
